+++ business-easy-core/src/main/java/com/businesseasy/core/repositories/PeopleRepository.java
@@ -8,9 +8,9 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface PeopleRepository extends JpaRepository<PeopleEntity, Integer>, JpaSpecificationExecutor<PeopleEntity> {
+public interface PeopleRepository extends JpaRepository<PeopleEntity, Long>, JpaSpecificationExecutor<PeopleEntity> {
     PeopleEntity findByName(String name);
-    Optional<PeopleEntity> findById(Integer id);
+    Optional<PeopleEntity> findById(Long id);
 
     @Query(value = "SELECT p FROM PeopleEntity p JOIN FETCH p.contactNoList c " +
             " where p.type in ('BOTH', 'SUPPLIER')")
