+++ business-easy-core/src/main/java/com/businesseasy/core/services/people/PeopleService.java
@@ -22,5 +22,5 @@ public interface PeopleService {
 
     List<PeopleEntity> getAllSupplier();
 
-    People getPeopleById(Integer id);
+    People getPeopleById(Long id);
 }
